+++ keeper/relay.go
@@ -10,10 +10,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
-	clienttypes "github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
-	channeltypes "github.com/cosmos/ibc-go/v8/modules/core/04-channel/types"
-	host "github.com/cosmos/ibc-go/v8/modules/core/24-host"
-	coretypes "github.com/cosmos/ibc-go/v8/modules/core/types"
+	clienttypes "github.com/cosmos/ibc-go/v10/modules/core/02-client/types"
+	channeltypes "github.com/cosmos/ibc-go/v10/modules/core/04-channel/types"
 
 	"github.com/bianjieai/nft-transfer/types"
 )
@@ -65,13 +63,8 @@ func (k Keeper) SendTransfer(
 		return 0, errorsmod.Wrapf(channeltypes.ErrChannelNotFound, "port ID (%s) channel ID (%s)", sourcePort, sourceChannel)
 	}
 
-	destinationPort := channel.GetCounterparty().GetPortID()
-	destinationChannel := channel.GetCounterparty().GetChannelID()
-
-	channelCap, ok := k.scopedKeeper.GetCapability(ctx, host.ChannelCapabilityPath(sourcePort, sourceChannel))
-	if !ok {
-		return 0, errorsmod.Wrap(channeltypes.ErrChannelCapabilityNotFound, "module does not own channel capability")
-	}
+	destinationPort := channel.Counterparty.PortId
+	destinationChannel := channel.Counterparty.ChannelId
 
 	// See spec for this logic: https://github.com/cosmos/ibc/blob/master/spec/app/ics-721-nft-transfer/README.md#packet-relay
 	packet, err := k.createOutgoingPacket(ctx,
@@ -87,15 +80,15 @@ func (k Keeper) SendTransfer(
 		return 0, err
 	}
 
-	sequence, err := k.ics4Wrapper.SendPacket(ctx, channelCap, sourcePort, sourceChannel, timeoutHeight, timeoutTimestamp, packet.GetBytes())
+	sequence, err := k.ics4Wrapper.SendPacket(ctx, sourcePort, sourceChannel, timeoutHeight, timeoutTimestamp, packet.GetBytes())
 	if err != nil {
 		return 0, err
 	}
 
 	defer func() {
 		labels := []metrics.Label{
-			telemetry.NewLabel(coretypes.LabelDestinationPort, destinationPort),
-			telemetry.NewLabel(coretypes.LabelDestinationChannel, destinationChannel),
+			telemetry.NewLabel("destination_port", destinationPort),
+			telemetry.NewLabel("destination_channel", destinationChannel),
 		}
 
 		telemetry.SetGaugeWithLabels(